Store timestamp in s64, fixes double clicks, broken in 3c93bbc

+++ src/input/Input.h
@@ -163,7 +163,7 @@ private:
 	bool  bMouseButton[Mouse::ButtonCount];
 	bool  bOldMouseButton[Mouse::ButtonCount];
 	
-	int   iMouseTime[Mouse::ButtonCount];
+	s64   iMouseTime[Mouse::ButtonCount];
 	int   iMouseTimeSet[Mouse::ButtonCount];
 	int   iOldNumClick[Mouse::ButtonCount];
